\name{bullwhipgame}
\alias{bullwhipgame}
\title{Bullwhip effect demo in shiny}
\usage{
bullwhipgame()
}
\value{
  Nothing
}
\description{
bullwhipgame is a Shiny app to illustrate and visualize the bullwhip effect. The bullwhip effect is commonly defined as the ratio between the unconditional variance of the order process to that of the demand
process.
}
\examples{
## Not run:
#library(shiny)
#bullwhipgame()
}
